Implementa os serviços necessários para o processo de checkout
