+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/checkout/StartCheckoutService.java
@@ -1,4 +1,9 @@
 package br.com.fiap.controlepedidos.core.application.services.checkout;
 
+import br.com.fiap.controlepedidos.core.domain.entities.Payment;
+
+import java.util.UUID;
+
 public interface StartCheckoutService {
+    public Payment startCheckout(UUID cartId) throws Exception;
 }
