+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/checkout/impl/ConfirmQRCodePaymentOrderImpl.java
@@ -1,36 +1,42 @@
 package br.com.fiap.controlepedidos.core.application.services.checkout.impl;
 
-import br.com.fiap.controlepedidos.core.application.ports.IPaymentGateway;
-import br.com.fiap.controlepedidos.core.application.services.checkout.StartCheckoutService;
-import br.com.fiap.controlepedidos.core.application.services.order.CreateOrderService;
+import br.com.fiap.controlepedidos.core.application.services.checkout.ConfirmQRCodePaymentOrder;
+import br.com.fiap.controlepedidos.core.application.services.order.FindOrderByIdService;
+import br.com.fiap.controlepedidos.core.application.services.order.PayOrderService;
 import br.com.fiap.controlepedidos.core.domain.entities.Order;
 import br.com.fiap.controlepedidos.core.domain.entities.Payment;
+import br.com.fiap.controlepedidos.core.domain.enums.OrderStatus;
+import br.com.fiap.controlepedidos.core.domain.enums.PaymentStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-public class StartCheckoutServiceImpl implements StartCheckoutService {
+public class ConfirmQRCodePaymentOrderImpl implements ConfirmQRCodePaymentOrder {
 
-    private final CreateOrderService createOrderService;
-    private final IPaymentGateway paymentGateway;
+    private final FindOrderByIdService findOrderByIdService;
+    private final PayOrderService payOrderService;
 
-    public StartCheckoutServiceImpl(CreateOrderService createOrderService, IPaymentGateway paymentGateway) {
-        this.createOrderService = createOrderService;
-        this.paymentGateway = paymentGateway;
+    public ConfirmQRCodePaymentOrderImpl(FindOrderByIdService findOrderByIdService, PayOrderService payOrderService) {
+        this.findOrderByIdService = findOrderByIdService;
+        this.payOrderService = payOrderService;
     }
 
     @Override
-    public Payment startCheckout(UUID cartId) throws Exception {
+    public Payment confirmQrCodePayment(UUID orderId, float paidValue) throws Exception {
+        try {
+            Order orderToPay = findOrderByIdService.getById(orderId);
 
-        //Creating Order and Payment to Request Payment Gateway
-        Order newOrder = createOrderService.createOrder(cartId);
-        Payment payment = paymentGateway.generatePixQrCodeMercadoPago(newOrder);
+            Payment result = payOrderService.payOrder(orderToPay, paidValue);
 
-        //TODO realizar implementação de pagamento de payment no futuro
-        payment.setOrder(newOrder);
-        payment.setTotalCents(newOrder.getTotalCents());
-
-        return payment;
+            if (result.getPaymentStatus().equals(PaymentStatus.PAID)) {
+                return result;
+            } else {
+                //TODO ao implementar pagamento é necessário atualizar o pagamento aqui também.
+                return new Payment();
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 }
