+++ src/main/java/br/com/fiap/controlepedidos/core/application/services/order/FinishOrderService.java
@@ -4,6 +4,6 @@ import br.com.fiap.controlepedidos.core.domain.entities.Order;
 
 import java.util.UUID;
 
-public interface StartOrderPreparation {
-    Order startPreparation(UUID orderId) throws Exception;
+public interface FinishOrderService {
+    Order perform(UUID orderId) throws Exception;
 }
